Refactor import paths for addGdsBodyClass
