+++ src/app/components/govuk-accordion/govuk-accordion.component.spec.ts
@@ -1,15 +1,13 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { GovukAccordionComponent } from './govuk-accordion.component';
-import { addGdsBodyClass } from '../utils/add-gds-body-class';
+import { addGdsBodyClass } from '../helpers/index';
 
 describe('GovukAccordionComponent', () => {
   let component: GovukAccordionComponent;
   let fixture: ComponentFixture<GovukAccordionComponent>;
 
-  // We need to add the govuk-frontend-supported class to the body element as per govuk-frontend v5
-  addGdsBodyClass();
-
+  beforeAll(addGdsBodyClass);
   beforeEach(() => {
     TestBed.configureTestingModule({
       imports: [GovukAccordionComponent],
