+++ src/app/components/helpers/index.ts
@@ -1,0 +1,3 @@
+import { addGdsBodyClass } from './add-gds-body-class';
+
+export { addGdsBodyClass };
