--- src/app/components/utils/add-gds-body-class.ts
@@ -1,5 +1,0 @@
-export function addGdsBodyClass() {
-  if (!document.body.classList.contains('govuk-frontend-supported')) {
-    document.body.classList.add('govuk-frontend-supported');
-  }
-}
